Functioning aggregation from Varsom and RegObs
This module is somewhat complete now. More options are needed
fine-tuning what data goes into a LabeledData, however. ForecastDataset
should also be changed in the future to support RegObs and weather data
for days without an published avalanche bulletin.

The suggested workflow is to create a ForecastDataset once,
generate a number of different LabeledData from it and write
those to csv-files. To use the datasets later, use LabeledData.from_csv()
to read in from disk.

This way, the only recurring delay using the data is to read CSVs from disk,
instead of relying on the web services or pickled caches of data formatted
in an unsuitable manner.