Add distance measurement between forecasts
This is a way to compare avalanche bulletins. It is a kind of
distance measurement. It is commutative but does not fulfill the
triangle inequality, as the null forecast is defined as a plane.

It works in three stages. First it measures the difference between
avalanche problems. Then it measures the difference between ordered
lists of avalanche problems. This is called the p_score. The last
stage is to compare the tuple
(danger_level, emergency_warning, p_score).

It is currently rather unbalanced. This will be adjusted later.