Stops sk_clustered from dropping columns
Apparently pandas.DataFrame.mode() removes columns if they are
all np.nan. This poses a problem for LabeledData.f1(), since it
expects the same columns in LabeledData.label and LabeledData.pred.

This was discovered when training an ensemble model. It seems to
be relatively unusual for an entire column to be only np.nan in our
dataset.