Remove LabeledData.kfold() (#29)
Recent testing has shown that a split of a dataset looks to be
a lot more like the other partition of the dataset than a separate
season. It is therefore unsuitable to use for testing.

A (untested) theory to why this might be is that the forecasts
are spatially autocorrelated, and that neighbouring regions with
similar data will end up in different partitions, _especially_
when stratifying the data.