Meta model model (#33)
* Set max_file_age=0

This should not have any practical effect.

* Meta model

* Handle LabeledData without label

* Handle temperature stretching after loading the LabeledData

* Update submodules

* Create full meta model with REAL and MULTI values

* Stops sk_clustered from dropping columns

Apparently pandas.DataFrame.mode() removes columns if they are
all np.nan. This poses a problem for LabeledData.f1(), since it
expects the same columns in LabeledData.label and LabeledData.pred.

This was discovered when training an ensemble model. It seems to
be relatively unusual for an entire column to be only np.nan in our
dataset.

* Updated meta model

* Handle 'force_subprobs' correctly

* Fix errors in LabeledData.valid_pred()