Update to Regobs API v4
Updated to be able to use the Observation class in varsomdata.

A lot of new observations use lists to accomodate for multiple
observations of the same type in one registration. Currently,
we only use the first observation in such a list for compability.
This may be changed in the future.